Separated functions to files :recycle:
Relates to #1233
